refactor evals structure
