+++ src/leapfrogai_evals/evals/__init__.py
@@ -1,0 +1,7 @@
+# __init__.py
+# ruff: noqa: F401
+
+from leapfrogai_evals.evals.human_eval import human_eval
+from leapfrogai_evals.evals.mmlu import mmlu
+from leapfrogai_evals.evals.niah_eval import niah_eval
+from leapfrogai_evals.evals.qa_eval import qa_eval
