+++ src/leapfrogai_evals/models/__init__.py
@@ -1,0 +1,5 @@
+# __init__.py
+# ruff: noqa: F401
+
+from leapfrogai_evals.models.claude_sonnet import ClaudeSonnet
+from leapfrogai_evals.models.lfai import LFAI_Model
