+++ src/leapfrogai_evals/pyproject.toml
@@ -7,7 +7,7 @@ version = "0.13.0"
 # x-release-please-end
 
 dependencies = [
-    "deepeval == 1.2.2",
+    "deepeval == 1.3.0",
     "openai == 1.42.0",
     "tqdm == 4.66.5",
     "python-dotenv == 1.0.1",
