+++ src/leapfrogai_evals/runners/__init__.py
@@ -1,0 +1,5 @@
+# __init__.py
+# ruff: noqa: F401
+
+from leapfrogai_evals.runners.niah_runner import NIAH_Runner
+from leapfrogai_evals.runners.qa_runner import QA_Runner
